invoice
invoice

+++ backend/App/controllers/invoice-controller.js
@@ -44,20 +44,28 @@ invoicesCltr.create=async(req,res)=>{
         res.status(500).json({errors:'Internal server errrors'})
     }
 }
-invoicesCltr.list=async(req,res)=>{
+invoicesCltr.list = async(req,res)=>{
     try{
-        const invoice=await Invoices.find()
-        res.json(invoice)
+        const id = req.params.id
+        const invoice = await Invoices.findOne({_id:id})
+    if(!invoice){
+        return res.status(404).json({ error: 'Invoice not found' });
+    }
+    res.json(invoice)
     }catch(err){
-        console.log(err)
-        res.status(201).json({errors:'Internal server errors'})
+      console.log(err)
+      res.status(500).json({error:'Internal Server Error'})
     }
 }
 
 invoicesCltr.delete=async(req,res)=>{
     try{
         const id= req.params.id
-        const invoice=await Invoices.findOneAndDelete({_id:id})
+        const userId = req.user.id
+        const invoice=await Invoices.findOneAndDelete({_id:id, userId : userId})
+        if (!invoice) {
+            return res.status(404).json({ error: 'Invoice not found' });
+        }
         res.status(200).json(invoice)
     }
     catch(err){
